# src/main/resources/doc_swagger.yaml
swagger: '2.0'
info:
  description: >-
    Sugestão de documentação para ser utilizada como base para o desenvolvimento
    da API do projeto OPEN PMO da SEP-ES. 
  version: 1.0.0
  title: Open PMO Api
  termsOfService: 'http://www.totemti.com.br/'
  contact:
    email: atendimento@totemti.com.br
host: localhost
basePath: /openPmo
...
schemes:
  - https
  - http
paths:
...
  /plan-permissions:
    get:
      tags:
        - planPermissionController
      summary: Get all plan permissions
      description: ''
      operationId: getAllPlanPermission
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: id-plan
          description: id-plan
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/PlanPermissionDto'
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    post:
      tags:
        - planPermissionController
      summary: Add a new plan permissions
      description: ''
      operationId: postPermissionPlan
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/PlanPermissionParamDto'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EntityDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    put:
      tags:
        - planPermissionController
      summary: Update an plan Permission
      description: ''
      operationId: planPermissionUpdate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/PlanPermissionParamDto'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
  '/workpacks/{id}':
    get:
      tags:
        - workpackController
      summary: Get workpack from a plan
      description: ''
      operationId: getWorkpack
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from workpack
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/WorkpackDetailDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    delete:
      tags:
        - workpackController
      summary: Delete a workpack from a plan
      description: ''
      operationId: deleteWorkpack
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from workpack
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
  /schedules:
    get:
      tags:
        - scheduleController
      summary: Get all workpack from a plan
      description: ''
      operationId: getAllSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: id-workpack
          description: id from workpack
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/ScheduleDto'
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    post:
      tags:
        - scheduleController
      summary: Add a new workpack to a plan
      description: ''
      operationId: addSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: id-workpack
          description: id from workpack
          required: true
          type: integer
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/ScheduleParamDto'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/EntityDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    put:
      tags:
        - scheduleController
      summary: Update a schedule from a workpack
      description: ''
      operationId: updateSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/ScheduleParamDto'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error          
  '/schedules/{id}':
    get:
      tags:
        - scheduleController
      summary: Get a schedule from a workpack
      description: ''
      operationId: getSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from schedule
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/ScheduleDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    delete:
      tags:
        - scheduleController
      summary: Delete a schedule from a workpack
      description: ''
      operationId: deleteSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from schedule
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
  /organizations:
    get:
      tags:
        - organizationController
      summary: Get all organizations
      description: ''
      operationId: getAllOrganization
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/OrganizationDto'
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    post:
      tags:
        - organizationController
      summary: Add a new organization
      description: ''
      operationId: addOrganization
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/OrganizationParamDto'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/EntityDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    put:
      tags:
        - organizationController
      summary: Update an organization
      description: ''
      operationId: updateOrganization
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/OrganizationParamDto'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
  '/organizations/{id}':
    get:
      tags:
        - organizationController
      summary: Get an organization
      description: ''
      operationId: getOrganization
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/OrganizationDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    delete:
      tags:
        - organizationController
      summary: Delete an organization
      description: ''
      operationId: deleteOrganization
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  EntityDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
...
  PlanPermissionDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idPerson:
        type: integer
        format: int64
      permissionLevel:
        type: string
      permittedRole:
        type: string
      organization:
        type: string
  PlanPermissionParamDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idPlan:
        type: integer
        format: int64
      idPerson:
        type: integer
        format: int64
      permissionLevel:
        type: string
      permittedRole:
        type: string
      organization:
        type: string
...
  WorkpackDetailDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack
        enum:
          - Portfolio
          - Program
          - Organizer
          - Deliverable
          - Project
          - Milestone
      workpackModel:
        $ref: '#/definitions/WorkpackModelDetailDto'
      properties:
        type: array
        items:
          $ref: '#/definitions/PropertyDto'
      children:
        type: array
        items:
          $ref: '#/definitions/WorkpackModelDto'
...
  PropertyDto:
    type: object
    discriminator: type
    required:
      - type
      - idPropertyModel
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - Integer
          - Text
          - Date
          - Toggle
          - UnitSelection
          - Selection
          - TextArea
          - Number
          - Currency
          - LocalitySelection
          - OrganizationSelection
      idPropertyModel:
        type: integer
        format: int64
...
  WorkpackModelDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - PortfolioModel
          - ProgramModel
          - OrganizerModel
          - DeliverableModel
          - ProjectModel
          - MilestoneModel
      modelName:
        type: string
      fontIcon:
        type: string
...
  WorkpackModelDetailDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idPlanModel:
        type: integer
        format: int64
      idParent:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - PortfolioModel
          - ProgramModel
          - OrganizerModel
          - DeliverableModel
          - ProjectModel
          - MilestoneModel
      modelName:
        type: string
      modelNameInPlural:
        type: string
      fontIcon:
        type: string
      sortBy:
        type: string
      costSessionActive:
        type: boolean
      stakeholderSessionActive:
        type: boolean
      scheduleSessionActive:
        type: boolean
      childWorkpackModelSessionActive:
        type: boolean
      personRoles:
        type: array
        items:
          type: string
      organizationRoles:
        type: array
        items:
          type: string
      properties:
        type: array
        items:
          $ref: '#/definitions/PropertyModelDto'
...
  PropertyModelDto:
    type: object
    discriminator: type
    required:
      - type
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - IntegerModel
          - TextModel
          - DateModel
          - ToggleModel
          - UnitSelectionModel
          - SelectionModel
          - TextAreaModel
          - NumberModel
          - CurrencyModel
          - LocalitySelectionModel
          - OrganizationSelectionModel
      name:
        type: string
      label:
        type: string
      sortIndex:
        type: integer
      fullLine:
        type: boolean
      required:
        type: boolean
...
  OrganizationDto:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      fullName:
        type: string
      phoneNumber:
        type: string
      address:
        type: string
      contactEmail:
        type: string
      website:
        type: string
      sector:
        type: string
        enum:
          - public
          - private
          - third
  OrganizationParamDto:
    type: object
    required:
      - name
      - fullName
      - phoneNumber
      - address
      - contactEmail
      - sector
    properties:
      name:
        type: string
      fullName:
        type: string
      phoneNumber:
        type: string
      address:
        type: string
      contactEmail:
        type: string
      website:
        type: string
      sector:
        type: string
        enum:
          - public
          - private
          - third
...
  ScheduleDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      start:
        description: Date of pattern yyyy/MM/dd
        type: string
      finish:
        description: Date of pattern yyyy/MM/dd
        type: string
      plannedWork:
        type: number
      actualWork:
        type: number
      costs:
        type: array
        items:
          $ref: '#/definitions/StepCostParamDto'
  ScheduleParamDto:
    type: object
    required:
      - start
      - finish
      - plannedWork
    properties:
      start:
        description: Date of pattern yyyy/MM/dd
        type: string
      idWorkpack:
        type: integer
        format: int64
      finish:
        description: Date of pattern yyyy/MM/dd
        type: string
      plannedWork:
        type: number
      actualWork:
        type: number
      costs:
        type: array
        items:
          $ref: '#/definitions/StepCostParamDto'
...
  StepCostParamDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      plannedCost:
        type: number
      actualCost:
        type: number
...